refactor: extract YouTube integration from docs and convert to TS
