+++ packages/docs/package.json
@@ -17,6 +17,7 @@
     "@material-ui/core": "^4.11.2",
     "@mdx-js/react": "^1.6.21",
     "@techboi/consent-manager": "*",
+    "@techboi/consent-manager-integration-youtube": "*",
     "clsx": "^1.1.1",
     "final-form": "^4.20.1",
     "react": "^16.8.4",
